LafferCurve() to "helper_functions"
I've added a function that needs looking at. I've ensure the substance of the code works, but I need help creating the output directory and deciding how to best define the function. I think we'd like people to be able to call LafferCurve() or LafferCurve(2035) if they want to chang the year and always have the directory be fixed.