Add code to do a sensitivity run with FAIR climate model
This adds several functions to the "helper_functions.jl" file that (i) create an instance of FAIR to use in the optimization and (ii) create an objective function that iterates between the FAIR temperature projections and NICE endogenous emissions. It also adds code to carry out the actual optimzation in the paper submission replication file.